Create 1-D_Frogger_Easy.py
Description: This code takes an integer 'n', a starting position 'p', and a 'magic' number as input. It also takes a list of 'n' integers as input. It simulates a game where you start at position 'p' in the list and move based on the values in the list until a certain condition is met.